income and budget tests

+++ models/__init__.py
@@ -65,7 +65,7 @@ class Income(db.Model):
     """Represents a single income record."""
     id = db.Column(db.Integer, primary_key=True)
     source = db.Column(db.String(50), nullable=False)
-    amount = db.Column(db.Float, nullable=False)
+    amount = db.Column(db.Numeric(10, 2), nullable=False)
     date = db.Column(db.Date, nullable=False)
     description = db.Column(db.Text, nullable=True)
